rename packages

+++ test/test.sh
@@ -1,3 +1,16 @@
+#! /bin/bash 
+###########################################
+#
+###########################################
+
+# constants
+baseDir=$(cd `dirname "$0"`;pwd)
+# functions
+
+# main 
+[ -z "${BASH_SOURCE[0]}" -o "${BASH_SOURCE[0]}" = "$0" ] || return
+cd $baseDir
 rm -rf gen-py
 thrift --gen py demo.thrift
 python demo.py
+
